More like a Gem
* Used a little trick in the bin to make it so it will load the lib path.
* Used the Chef class to surround everything
* Added a PullRequestCalculator class and added all that free-floating code to it.
